Refactored SPI HAL API to be more generic.
The location concept was specific to Gecko MCU's. For STM32 for instance possible SPI pins
are not limited to a specific configuration from a set of locations.
The SPI API now takes an SPI port number to used, where the ports are defined in the platform.
This way we can define the pins to use for each port per platform.
While doing this, the pin_id_t struct was removed as well, this is now a uint32_t.

Currently only stm32l0xx chip and NUCLEO_L073RZ is updated, other chips and platforms will follow.